feat: Enhance timestamp handling and add onboard power sampler
- Updated Sample struct in sample_queue.h to include new timestamp fields for device and host.
- Removed device timestamp overflow handling from session.h and session.cpp.
- Modified main.cpp to add debug flag for time synchronization diagnostics.
- Improved power_monitor_session.cpp to handle new timestamp formats and added time sync debug logging.
- Updated read_thread.cpp to support 64-bit timestamps and adjusted data sample size.
- Added onboard power sampler to read and log power metrics from hwmon interface with configurable parameters.
- Enhanced functional tests to verify time synchronization and data integrity under various conditions.

+++ device/state_machine.hpp
@@ -31,7 +31,7 @@ struct DeviceContext {
     // Streaming configuration
     uint16_t stream_period_us = 1000;   // Default 1kHz
     uint16_t stream_mask = 0x000F;      // All channels enabled
-    uint32_t stream_start_us = 0;       // Timestamp when streaming started
+    uint64_t stream_start_us = 0;       // Timestamp when streaming started
     bool usb_stress_mode = false;       // Continuous fixed-sample streaming for USB throughput test
 
     // INA228 configuration (from SET_CFG)
@@ -63,7 +63,7 @@ struct DeviceContext {
     }
 
     // Start streaming and update shared context for sampler core
-    void start_streaming(uint16_t period_us, uint16_t mask, uint32_t now_us) {
+    void start_streaming(uint16_t period_us, uint16_t mask, uint64_t now_us) {
         stream_period_us = period_us;
         stream_mask = mask;
         stream_start_us = now_us;
